sep-16-2018-consolidation
