+++ src/main/java/axisallies/board/BoardBuilder.java
@@ -25,6 +25,18 @@ public class BoardBuilder {
     private static final String USSR_LAND_3_TERRITORIES_JSON = "ussr_land_3_territories.json";
     private static final String TEST_PLAYER_SETUP = "sample_players.json";
 
+    public static Board testBuild() throws IOException {
+        return build(UnitType.class, USSR_LAND_3_TERRITORIES_JSON, TEST_PLAYER_SETUP);
+    }
+
+    public static Board testBuild(String testTerritoryMap, String testPlayerSetupFile) throws IOException {
+        return build(UnitType.class, testTerritoryMap, testPlayerSetupFile);
+    }
+
+    public static Board sourceBuild() throws IOException {
+        return build(BoardBuilder.class, GAME_TERRITORY_MAP_SETUP, PLAYER_SETUP_FILE);
+    }
+
     private static void createTerritoryMap(Board board, String boardGameSetupFile, Class clazz) throws IOException {
 
         String jsonFilePath = getResourcePath(boardGameSetupFile, clazz);
@@ -81,18 +93,6 @@ public class BoardBuilder {
     private static String getResourcePath(String fileName, Class clazz) {
         ClassLoader classLoader = clazz.getClassLoader();
         return classLoader.getResource(fileName).getPath();
-    }
-
-    public static Board testBuild() throws IOException {
-        return build(UnitType.class, USSR_LAND_3_TERRITORIES_JSON, TEST_PLAYER_SETUP);
-    }
-
-    public static Board testBuild(String testTerritoryMap, String testPlayerSetupFile) throws IOException {
-        return build(UnitType.class, testTerritoryMap, testPlayerSetupFile);
-    }
-
-    public static Board sourceBuild() throws IOException {
-        return build(BoardBuilder.class, GAME_TERRITORY_MAP_SETUP, PLAYER_SETUP_FILE);
     }
 
     private static Board build(Class clazz, String mapFile, String playerFile) throws IOException {
