+++ src/main/java/axisallies/validators/MoveValidator.java
@@ -2,7 +2,7 @@ package axisallies.validators;
 
 import axisallies.board.Territory;
 import axisallies.nations.NationType;
-import axisallies.units.Path;
+import axisallies.board.Path;
 import axisallies.units.Unit;
 
 import java.util.Collection;
